requirements.txt
accelerate==0.28.0
datasets==2.18.0
evaluate==0.4.1
gensim==4.3.0
googletrans==3.1.0a0
h5py==3.10.0
huggingface_hub==0.21.4
jieba==0.42.1
matplotlib==3.7.4
nltk==3.8.1
openai==0.28.0
paddlepaddle==2.4.2
pandas==2.0.3
python_Levenshtein==0.25.1
rankgen==0.1.1
Requests==2.31.0
retriv==0.2.3
sacremoses==0.1.1
scikit_learn==1.3.2
seaborn==0.13.2
sentence_transformers==2.5.1
sentencepiece==0.2.0
setuptools==69.2.0
simcse==0.4
six==1.14.0
spacy==3.7.4
torch==2.2.2
tqdm==4.66.1
wandb==0.16.6
